requirements.txt
yt-dlp
librosa
torch
openai-whisper
numpy